+++ adlfs/tests/test_uri_format.py
@@ -34,6 +34,7 @@ def test_dask_parquet(storage):
             "{}://test@dfs.core.windows.net/test_group.parquet".format(protocol),
             storage_options=STORAGE_OPTIONS,
             engine="pyarrow",
+            write_metadata_file=True,
         )
 
         fs = AzureBlobFileSystem(**STORAGE_OPTIONS)
@@ -50,3 +51,10 @@ def test_dask_parquet(storage):
         engine="pyarrow",
     ).compute()
     assert_frame_equal(df, df_test)
+
+
+def test_account_name_from_url():
+    kwargs = AzureBlobFileSystem._get_kwargs_from_urls(
+        "abfs://test@some_account_name.dfs.core.windows.net/some_file"
+    )
+    assert kwargs["account_name"] == "some_account_name"