+++ CHANGELOG.md
@@ -1,4 +1,24 @@
 **Change Log**
+2022.10.1
+----------
+- Reorder fs.info() to search the parent directory only after searching for the specified item directly
+- Pin azure-storage-blob >=12.12.0,<12.14.  Requires pinning azure-core>=1.23.1,<2.0.0
+
+2022.9.1
+--------
+- Fixed missing dependency on `aiohttp` in package metadata.
+
+2022.9.0
+--------
+- Add support to AzureBlobFileSystem for versioning
+- Assure full uri's are left stripped to remove "\"
+- Set Python requires >=3.8 in setup.cfg
+- _strip_protocol handle lists
+
+2022.7.0
+--------
+- Fix overflow error when uploading files > 2GB
+
 2022.04.0
 ---------
 - Added support for Python 3.10 and pinned Python 3.8